requirements.txt
Django==1.11.6
dj-database-url==0.4.2
django-fsm==2.6.0
django-push-notifications==1.5.0

djangorestframework==3.7.1

psycopg2==2.7.3.1
requests==2.18.4
